dev-lang/rust: upgrade to rust 1.25
The C/C++ compiler was changed for the x86-64 target to
x86_64-pc-linux-gnu-clang/clang++ because the internal LLVM
requires libxml2 to be installed and findable by the compiler.

There are shared objects that need to be installed from a different
directory in the build root that was added by this ebuild.

BUG=None
TEST=sudo emerge rust

Change-Id: I89a647653b0bf49b583f96601764a889d09338fa
Reviewed-on: https://chromium-review.googlesource.com/1013206
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>