linux-headers: upgrade nl80211.h
We may want to start using NL80211_ATTR_SCHED_SCAN_MAX_REQS in shill, so
let's update our nl80211 header. Of couse, we'll keep compatibility if
the attribute is missing, but we have to at least have the definition if
we want programs to compile.

This diff syncs with Linux v4.16.

BUG=b:72944404
TEST=trybots

Change-Id: Ida901bac1ce824cec2f646a4db1c96bd750dc657
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1014479
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>