libchrome: Move preg_parser from libchrome into authpolicy
Moves the .cc, .h, unit test + testdata and fuzzer for preg_parser into
authpolicy/policy, out from libchrome. This is because preg_parser is no
longer used by chrome or libchrome, authpolicy is the only remaining
user.

(This is the libchrome half of a change that touches libchrome and
authpolicy.)

BUG=chromium:832044
TEST=Build libchrome and inspect it to see it doesn't have preg_parser any more.
CQ-DEPEND=CL:1016605

Change-Id: I1cbbd9f3212a97ded069ffce11d03c0dff04cdef
Reviewed-on: https://chromium-review.googlesource.com/1016763
Commit-Ready: A Olsen <olsen@chromium.org>
Tested-by: A Olsen <olsen@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>