strongswan: allow spawning as non-root user.
This patch allows for giving strongSwan only the runtime capabilities it
needs, rather than full root privileges.

Adds preprocessor directives which allow strongSwan to be configured to
 1) start up as a non-root user
 2) avoid modprobe()'ing IPsec kernel modules into the kernel, which
    would normally require root or CAP_SYS_MODULE

Additionally, some small mods to charon/libstrongswan ensure that charon
supports starting as a non-root user.

Tested with strongSwan 5.5.3.

This patch can go in as is without changing any functionality in CrOS,
and the change should be landing in upstream as well:
https://lists.strongswan.org/pipermail/dev/2018-April/001905.html.

BUG=chromium:649417
TEST=tested as part of larger shill sandboxing debug CL

Change-Id: I55a70cf8ab6dab809d3f1980ec1eac45def71d85
Reviewed-on: https://chromium-review.googlesource.com/1015966
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>