intel-hybrid-driver: respect wayland configure flags
The configure script for this package ignores the user request to
disable wayland if the WAYLAND_SCANNER m4 macro is defined.  This wasn't
a problem before since wayland wasn't installed in the sdk and so the
macro was not defined.

Now that we're actually trying to get wayland into the sdk this is
breaking the builders.  Fix the configure.ac file to respect the flags
that are given to it.

BUG=chromium:826923
TEST=build_packages --board zako

Change-Id: I478fec6066c17a47a225aafc428b9405642541f7
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1015984
Reviewed-by: David Reveman <reveman@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>