eclass/cros-kernel: Update x86_64_defconfig
With the existing x86_64_defconfig, ChromeOS does not boot to GUI
when building the upstream kernel. It does not instantiate network
interfaces. Use configuration based on chromeos-4.14 to update it.
The resulting image boots v4.17-{rc1,rc2} to GUI, with at least sometimes
functional network interfaces, keyboard, mouse, and touchscreen.

Change-Id: I5579b880864637f42ca7e0d5cb01ee9d89f19009
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1024800
Reviewed-by: Mike Frysinger <vapier@chromium.org>