openvpn: remove openvpn-2.4.1-iv_plat.patch
This change is in preparation to run the shill process tree as a
non-root user.

Currently, CrOS carries forward an openvpn patch that uses environment
variables to set the CrOS platform and release version info that will be
pushed to the VPN server. Abandon this patch since we will soon be
running openvpn in a minijail, which doesn't support the setting of
arbitrary environment variables. Rather than use environment variables,
we will use the command line arg '--setenv NAME VALUE' to push info to
the server. This is done in chromium:999738.

CQ-DEPEND=CL:999738
BUG=chromium:649417
TEST=doesn't affect client side VPN operation, these values are only for
logging on the server side. see b/78029510 for details.

Change-Id: Ibb9ddd57d9a4dba29484ed63464983a35877d778
Reviewed-on: https://chromium-review.googlesource.com/1022177
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>