libchrome: Don't DCHECK unexpected message type, but ignore it
In dbus::ExportedObject, an assumption was made that no signal could be
received for an exported object, so DCHECK was put there. Now that there
is a use case of D-Bus forwarding by Bluetooth dispatcher, it is fine
that an exported object receives D-Bus signal messages as it is a
property change signal that is handled by dbus::ObjectManager. So we can
just ignore message types other than method in dbus::ExportedObject
rather than DCHECK it.

This is a cherry-pick of CL:1027081.

BUG=chromium:812468,chromium:835457
TEST=Build and platform2/bluetooth no longer crashes

Change-Id: I2dfc7c195bc5e72d2760e923d3f62174abc3fdfc
Reviewed-on: https://chromium-review.googlesource.com/1028712
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>